docs,refactor(evm): more function docs and better code organization
